split extension.js into modular structure for maintainability
