+++ Makefile
@@ -24,5 +24,5 @@ pack: compile-po ## Create a ZIP package for Extensions.gnome.org
 	@printf 'Packaging extension...\n'
 	@rm -f $(UUID).zip
 	@glib-compile-schemas schemas
-	@zip -qr $(UUID).zip extension.js prefs.js anims.js metadata.json stylesheet.css LICENSE schemas locale -x "schemas/gschemas.compiled" -x "po/generate.py"
+	@zip -qr $(UUID).zip *.js metadata.json stylesheet.css LICENSE schemas locale -x "schemas/gschemas.compiled" -x "po/generate.py"
 	@printf 'Created package: %s\n' "$(UUID).zip"
